% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{aggregate_data}
\alias{aggregate_data}
\title{Aggregate birth-level data into mother-level data}
\usage{
aggregate_data(birth_level_data)
}
\arguments{
\item{birth_level_data}{the dataset to aggregate}
}
\value{
a \code{tibble}
}
\description{
This function aggregates the data so that each row corresponds to a single mother.
}
\examples{
aggregate_data(data_births_all)

}
