% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{build_data_summary.table}
\alias{build_data_summary.table}
\title{Create data summary tables}
\usage{
build_data_summary.table(birth_level_data)
}
\arguments{
\item{birth_level_data}{the dataset to summarize}
}
\value{
a \code{tibble}
}
\description{
This function is used to create the tables providing the details of the data used in the study.
}
\examples{
build_data_summary.table(data_births_all)

}
