% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{build_fit_summary.table}
\alias{build_fit_summary.table}
\title{Create fit summary tables}
\usage{
build_fit_summary.table(fit)
}
\arguments{
\item{fit}{the output of one of the function fitting models (see \code{\link{fit_models}})}
}
\value{
a \code{tibble} with an additional attribute called "formula"
}
\description{
This function is used to create the tables providing the details of the fitted models used in the study.
}
\examples{
#See ?twinR

}
