% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI.R
\name{build_SI}
\alias{build_SI}
\title{Build Supplementary Information}
\usage{
build_SI(path_SI = "SI", overwrite = FALSE)
}
\arguments{
\item{path_SI}{the computer path to the folder where the pdf of the compiled SI will be written}

\item{overwrite}{a boolean indicating whether or not to overwrite a potentially existing SI.pdf file (default = FALSE)}
}
\description{
This function compiles the SI. It relies on a template "SI.Rnw" located in the package folder
\code{extdata}. For the compilation to work, you need to have created all the SI figures and tables,
and you need your system to be properly setup for the creation of LaTeX-PDF using knitr.
}
\examples{
# See ?twinR

}
