% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{build_summary_table}
\alias{build_summary_table}
\title{Create data summary tables}
\usage{
build_summary_table(data)
}
\arguments{
\item{data}{the dataset to summarize}
}
\value{
a \code{tibble}
}
\description{
This function is used to create the tables providing the details of the data used in the study.
}
\examples{
build_summary_table(data_births_all)

}
