% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit.R
\name{combine_simulated_slopes}
\alias{combine_simulated_slopes}
\title{Combine the simulated slopes across all scenarios}
\usage{
combine_simulated_slopes(path_slopes = "slopes_under_scenarios")
}
\arguments{
\item{path_slopes}{the path toward the folder where the objects containing the slopes have been
saved}
}
\value{
a tibble with all simulated slopes
}
\description{
Combine the simulated slopes across all scenarios
}
\examples{
#See ?twinR

}
