% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit.R
\name{computing_time_analysis}
\alias{computing_time_analysis}
\title{Extract computing time from the goodness-of-fit analysis}
\usage{
computing_time_analysis(slopes_obj)
}
\arguments{
\item{slopes_obj}{an object returned by \code{\link{simulate_slopes_for_GOF}} or \code{\link{combine_simulated_slopes}}}
}
\value{
a tibble with the information about the computing time in hours
}
\description{
This functions reads and summarizes the information about computing time contains in the objects created with \code{\link{simulate_slopes_for_GOF}} and \code{\link{combine_simulated_slopes}}.
}
\examples{
# See ?twinR

}
\seealso{
\code{\link{simulate_slopes_for_GOF}}, \code{\link{combine_simulated_slopes}}
}
