% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_births_all}
\alias{data_births_all}
\title{The complete dataset at the level of births}
\format{
a \code{tibble}
}
\description{
This dataset contains the raw data associated with 116,082 birth events.
It contains the following variables:
\itemize{
\item \code{pop} a factor with 9 levels defining the population
\item \code{maternal_id} a factor with 23,281 levels defining the identity of the mothers
\item \code{maternal_birthyear} the year when the mother was born
\item \code{maternal_age} the age of the mother (in months)
\item \code{birth_year} the year when a given birth occurred
\item \code{twin} the outcome of the birth in terms of twinning (\code{TRUE} or \code{FALSE})
\item \code{monthly} whether the observations were collected at a monthly resolution (\code{TRUE} or \code{FALSE})
}
}
\examples{
data_births_all

}
\references{
this study
}
\keyword{datasets}
