% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Display welcome message}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{argument needed but automatically defined}

\item{pkgname}{argument needed but automatically defined}
}
\description{
This function should not be called by the user.
It displays a message when the package is being loaded.
}
