% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{expand_data}
\alias{expand_data}
\title{Create derived variables for birth-level data}
\usage{
expand_data(birth_level_data)
}
\arguments{
\item{birth_level_data}{the dataset to expand}
}
\value{
a \code{tibble}
}
\description{
This function adds new columns to the birth level data, i.e.:
\itemize{
\item \code{age} the age of the mother at each birth (in years)
\item \code{parity} the rank of a given birth within the life history of a mother
\item \code{PP} whether the mother go on reproducing after a given birth
\item \code{IBI} the interbirth interval between a given birth and the next one (in months)
}
}
\details{
These new columns are derived from existing columns and are required for the
fit of statistical models.
}
\examples{
expand_data(data_births_all)

}
