% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{export_table_xlsx}
\alias{export_table_xlsx}
\title{Export a table in the *.xlsx format}
\usage{
export_table_xlsx(table, file)
}
\arguments{
\item{table}{a \code{tibble} or \code{data.frame} to be exported}

\item{file}{a string of characters indicating the name file to create (with path and extension required)}
}
\value{
the table (but the return is invisible)
}
\description{
This function is used to export tables into Microsoft Excel format.
It is a wrapper for the function \code{\link[openxlsx:write.xlsx]{write.xlsx}}.
}
\examples{
# See ?twinR

}
