% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{extract_poly.order}
\alias{extract_poly.order}
\title{Internal function for extracting the polynomial order of fitted models}
\usage{
extract_poly.order(fit)
}
\arguments{
\item{fit}{a fitted model}
}
\description{
This internal function allows for identifying which polynomial order has actually been used in a
fit. This is used for debugging purposes only.
}
