% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{figures}
\alias{figures}
\alias{draw_fig_1A}
\alias{draw_fig_1B}
\alias{draw_fig_1C}
\alias{draw_fig_1D}
\alias{draw_fig_2}
\alias{draw_fig_3A}
\alias{draw_fig_3B}
\alias{draw_fig_3C}
\alias{draw_fig_4}
\alias{draw_fig_S1}
\alias{draw_fig_S2}
\alias{draw_fig_S5}
\alias{draw_fig_S6A}
\alias{draw_fig_S6B}
\alias{draw_fig_S6C}
\alias{draw_fig_S6D}
\title{Functions producing the figures}
\usage{
draw_fig_1A(data)

draw_fig_1B(data)

draw_fig_1C(data)

draw_fig_1D(data)

draw_fig_2(data)

draw_fig_3A(data)

draw_fig_3B(data)

draw_fig_3C(data)

draw_fig_4(data)

draw_fig_S1(data)

draw_fig_S2(data)

draw_fig_S5(data)

draw_fig_S6A(data)

draw_fig_S6B(data)

draw_fig_S6C(data)

draw_fig_S6D(data)
}
\arguments{
\item{data}{a \code{data.frame} or a \code{list} containing the data to be plotted}
}
\value{
a ggplot
}
\description{
Functions producing the figures
}
\section{Functions}{
\itemize{
\item \code{draw_fig_1A}: draw fig. 1A

\item \code{draw_fig_1B}: draw fig. 1B

\item \code{draw_fig_1C}: draw fig. 1C

\item \code{draw_fig_1D}: draw fig. 1D

\item \code{draw_fig_2}: draw fig. 2

\item \code{draw_fig_3A}: draw fig. 3A

\item \code{draw_fig_3B}: draw fig. 3B

\item \code{draw_fig_3C}: draw fig. 3C

\item \code{draw_fig_4}: draw fig. 4

\item \code{draw_fig_S1}: draw fig. S1

\item \code{draw_fig_S2}: draw fig. S2

\item \code{draw_fig_S5}: draw fig. S5

\item \code{draw_fig_S6A}: draw fig. S6A

\item \code{draw_fig_S6B}: draw fig. S6B

\item \code{draw_fig_S6C}: draw fig. S6C

\item \code{draw_fig_S6D}: draw fig. S6D
}}

\examples{
# See ?twinR

}
