% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{filter_data}
\alias{filter_data}
\title{Filter the full data to retain full records}
\usage{
filter_data(birth_level_data)
}
\arguments{
\item{birth_level_data}{the dataset to filter}
}
\value{
a \code{tibble}
}
\description{
This function only keeps information from mothers with complete life history
and for which dates have been recorded at a monthly resolution.
}
\examples{
filter_data(data_births_all)

}
