% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_models}
\alias{fit_models}
\alias{fit_totalbirths}
\alias{fit_twinner.allbirths}
\alias{fit_twinner.firstbirth}
\alias{fit_twinning.prob}
\alias{fit_AFB}
\title{Fit the models}
\usage{
fit_totalbirths(
  mother_level_data,
  when_twinner = "allbirths",
  args_spaMM = list(),
  verbose = TRUE
)

fit_twinner.allbirths(mother_level_data, args_spaMM = list(), verbose = TRUE)

fit_twinner.firstbirth(mother_level_data, args_spaMM = list(), verbose = TRUE)

fit_twinning.prob(mother_level_data, args_spaMM = list(), verbose = FALSE)

fit_AFB(mother_level_data, args_spaMM = list(), verbose = FALSE)
}
\arguments{
\item{mother_level_data}{a \code{tibble} or \code{data.frame} with mother level data}

\item{when_twinner}{a string of characters indicating if the twinning status is based on all births ('allbirths') or just on the first one ('firstbirth')}

\item{args_spaMM}{list of additional arguments to pass to the function \code{\link[spaMM:fitme]{fitme}}}

\item{verbose}{whether to display the formula of the fit during the fitting procedure}
}
\value{
the fitted model
}
\description{
These functions fit the models used in the paper.
They have been programmed to be used with our specific data structure.
See \strong{Functions}, below, for details on each function.
}
\section{Functions}{
\itemize{
\item \code{fit_totalbirths}: fit the model predicting the total number of births per mother from her twinning status

\item \code{fit_twinner.allbirths}: fit the model predicting the twinning status of the mother at any birth during her life from her total number of births

\item \code{fit_twinner.firstbirth}: fit the model predicting the twinning status of the mother at first birth from her total number of births

\item \code{fit_twinning.prob}: fit the model predicting the probability of a birth to result in twins from the total number of births

\item \code{fit_AFB}: fit the model predicting the age at first birth from the twinning status and the total number of births
}}

\examples{
# See ?twinR

}
