% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{run_simulation}
\alias{run_simulation}
\title{Run a simulation}
\usage{
run_simulation(
  birth_level_data,
  scenario = NULL,
  life_history_fits = NULL,
  seed = 123L,
  output = list(birth_level_data.simulated = FALSE, slope = TRUE, fits = FALSE),
  timeout = Inf,
  verbose = list(fit = FALSE, simu = FALSE)
)
}
\arguments{
\item{birth_level_data}{a \code{tibble} or \code{data.frame} with birth level data (expanded or not)}

\item{scenario}{the scenario defining which models to be fitted: e.g. "ABCD", "AC"... (see paper
for explanations)}

\item{life_history_fits}{a named list containing the fits of each three life history models}

\item{seed}{an \code{integer} providing the seed for the random generator}

\item{output}{a named list of booleans with elements \code{birth_level_data.simulated}, \code{slope} and \code{fit} indicating what to output}

\item{timeout}{the maximal duration (in seconds) allowed for the fitting procedure (default =
Inf)}

\item{verbose}{whether to display the formula of the fit during the fitting procedure}
}
\value{
a list with the output requested
}
\description{
This function runs a single simulation. If the fits of the life history models are not provided,
they will be fitted within the function call.
}
\details{
The output is conditional on the input so to save memory footprint and only return whatever is
needed for a given application.
}
\examples{
# See ?twinR

}
\seealso{
\code{\link{simulate_slopes}}
}
