% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulate_reproduction}
\alias{simulate_reproduction}
\title{Simulate data to study the reproductive life of mothers}
\usage{
simulate_reproduction(
  birth_level_data,
  scenario,
  life_history_fits = NULL,
  effect_pt = 0,
  N_replicates = 100L,
  seed = 1L,
  nb_cores = 2L,
  lapply_pkg = "pbmcapply"
)
}
\arguments{
\item{birth_level_data}{a \code{tibble} or \code{data.frame} with birth level data (expanded or not)}

\item{scenario}{the scenario defining which models to be fitted: e.g. "ABCD", "AC"... (see paper
for explanations)}

\item{life_history_fits}{a named list containing the fits of each three life history models}

\item{effect_pt}{a numerical value defining by how much the intercept of \code{fit_pt} has to be
increased (on the scale of the linear predictor)}

\item{N_replicates}{the number of simulation replicates to run}

\item{seed}{an \code{integer} providing the seed for the random generator}

\item{nb_cores}{the number of CPU core(s) to use}

\item{lapply_pkg}{the R package used to implement a \code{lapply()} kind of function (default =
"pbmcapply"; other possibilities are "parallel" and "base")}
}
\value{
a tibble containing information about the reproductive life of the simulated mothers
}
\description{
This function simulates the life history of mothers and extract information about simulated
individuals. The argument \code{effect_pt} can be used to simulate an increased (or reduced) twinning
propensity.
}
\examples{
## See ?twinR

}
