% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit.R
\name{simulate_slopes}
\alias{simulate_slopes}
\title{Simulate slopes under a given scenario}
\usage{
simulate_slopes(
  birth_level_data,
  scenario,
  life_history_fits = NULL,
  seed = 123L,
  N_replicates = 49L,
  args_spaMM = list(),
  timeout = Inf,
  verbose = list(fit = FALSE, simu = FALSE)
)
}
\arguments{
\item{birth_level_data}{a \code{tibble} or \code{data.frame} with (expanded) birth level data}

\item{scenario}{the scenario defining which models to be fitted: e.g. "ABCD", "AC"... (see paper for explanations)}

\item{life_history_fits}{a named list containing the fits of each three life history models}

\item{seed}{an \code{integer} providing the seed for the random generator}

\item{N_replicates}{the number of simulation replicates to run}

\item{args_spaMM}{list of additional arguments to pass to the function \code{\link[spaMM:fitme]{fitme}}}

\item{timeout}{the maximal duration (in seconds) allowed for the fitting procedure (default = Inf)}

\item{verbose}{whether to display the formula of the fit during the fitting procedure}
}
\value{
a list containing the simulated slopes, the scenario, the input seed, and the time elapsed to do the job
}
\description{
This functions allows for the production of slopes of interest (the one quantifying effect of
total_birth on the log odd of twinning) taking a simulation scenario as true. This thus
corresponds to the computation of the slope under the null hypothesis (see Details for more
information). There is no reason to use this function directly. It is called by
\code{\link{simulate_slopes_for_GOF}}.
}
\examples{
## See ?twinR

}
\seealso{
simulate_slopes_for_GOF
}
