% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{test_parallel_computation}
\alias{test_parallel_computation}
\title{Testing parallel computing}
\usage{
test_parallel_computation(iter = 20L, nb_cores = 1L, list = NULL, cost = 1)
}
\arguments{
\item{iter}{the number of iteration to perform}

\item{nb_cores}{the number of CPU core(s) to use}

\item{list}{a list which length will be measured (optional)}

\item{cost}{the cost for the time threshold (default = 1; increase to speed up test, decrease to lengthen it)}
}
\description{
This function is aimed at testing if the parallel computation run efficiently and you can use it
as a skeleton to benchmark alternative implementations. It simply runs loops that take, in turns,
\code{(1:iter)/cost} time to run. A large object may be supplied to the argument \code{list} to monitor how
the memory is being handled. For example, you may want to provided to \code{list} the output of the
function \code{\link{fit_life_histories}}.
}
\details{
We tried many implementation (using the R packages parallel, furrr, future.apply and foreach;
combined with the backends doSnow, doParallel, or doFuture; using either multi-threading or
multi-processing). At least on our linux system, the implementation used here simply using
\code{\link[parallel:mclapply]{mclapply}} outperformed all these alternatives. The function must run under
a Unix based system and directly in a terminal (as opposed within R-GUI or RStudio) but it does
combine the different features:
\itemize{
\item no time seems wasted doing heavy handed communication between tasks
\item the handling of the memory is best: objects that can be shared between tasks are indeed shared and when a job
is done, the memory is immediately released
\item progress can be displayed (not in GUI)
}

And all of these nice features come without the need for any additional package!
}
\examples{
test_parallel_computation(iter = 4L, nb_cores = 2L)

}
