% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{theme_twin}
\alias{theme_twin}
\title{Theme for plots}
\usage{
theme_twin(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\value{
the theme
}
\description{
This is the custom ggplot theme that we use for all the figures produced in twinR.
The theme is built around \code{\link[ggplot2:ggtheme]{theme_bw}} and is inspired by
\code{\link[ggthemes:theme_few]{theme_few}}.
}
