% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinR.R
\docType{package}
\name{twinR-package}
\alias{twinR-package}
\alias{twinR}
\title{Welcome to the R package twinR}
\description{
This package contains the code necessary to produce the results of the paper
\emph{"Twinners in pre-industrial Europe: more babies but fewer births"}.
}
\details{
This document provides in its \strong{Examples} section the script producing all
the results of the paper.
}
\examples{

\dontrun{

#------------------------------------------------------------------------------------------------
#---------------------------------- Preparing datasets ------------------------------------------
#------------------------------------------------------------------------------------------------

## Filter the raw data to only keep data with monthly resolution:
data_births_monthly <- filter_data(data_births_all) # See ?filter_data

## Aggregate the data at the level of mothers:
data_mothers_all <- aggregate_data(data_births_all) # See ?aggregate_data
data_mothers_monthly <- aggregate_data(data_births_monthly)

## Expand the birth level data for the fit of statistical models:
data_births_monthly.complete <- expand_data(data_births_monthly) # See ?expand_data



#------------------------------------------------------------------------------------------------
#---------------------------------- Table 1 & S16: data summary ---------------------------------
#------------------------------------------------------------------------------------------------

dir.create("tables") # create a folder to store the tables

## Create table 1:
table1 <- build_summary_table(data_births_monthly)
table1
export_table_xlsx(table1, file = "tables/table1.xlsx")

## Create table S16:
tableS16 <- build_summary_table(data_births_all)
tableS16
export_table_xlsx(tableS16, file = "tables/tableS16.xlsx")



#------------------------------------------------------------------------------------------------
#---------------------------------- Fitting models ----------------------------------------------
#------------------------------------------------------------------------------------------------

fit_01 <- fit_totalbirths(data_mothers_monthly, when_twinner = "allbirths")
fit_02 <- fit_twinner.allbirths(data_mothers_monthly)
fit_03 <- fit_totalbirths(data_mothers_monthly, when_twinner = "firstbirth")
fit_04 <- fit_twinner.firstbirth(data_mothers_monthly)
fit_05 <- fit_twinning.binomial(data_mothers_monthly)
fit_05bis <- fit_twinning.binomial(data_mothers_all) # for Fig S5
fit_06 <- fit_AFB(data_mothers_monthly)
fit_07 <- fit_PP(data_births_monthly.complete, poly_order = 5L) # use best polynomial order
fit_08 <- fit_IBI(data_births_monthly.complete, poly_order = 6L)
fit_09 <- fit_twinning.binary(data_births_monthly.complete, poly_order = 3L)
fit_10 <- fit_PP(data_births_monthly.complete, twin_as.predictor = FALSE, poly_order = 5L)
fit_11 <- fit_IBI(data_births_monthly.complete, twin_as.predictor = FALSE, poly_order = 6L)
fit_12 <- fit_twinning.binary(data_births_monthly.complete) # no age and parity
fit_13 <- fit_twinning.binary(data_births_monthly.complete, maternal_ID_as.predictor = FALSE)
fit_14 <- fit_twinning.binary(data_births_monthly.complete, poly_order = 3L,
                              maternal_ID_as.predictor = FALSE)

}

}
\keyword{package}
