% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{figures}
\alias{figures}
\alias{draw_fig_1A}
\alias{draw_fig_1B}
\alias{draw_fig_1C}
\alias{draw_fig_1D}
\alias{draw_fig_2}
\alias{draw_fig_3A}
\alias{draw_fig_3B}
\alias{draw_fig_3C}
\title{Functions producing the figures}
\usage{
draw_fig_1A(data)

draw_fig_1B(data)

draw_fig_1C(data)

draw_fig_1D(data)

draw_fig_2(data)

draw_fig_3A(data)

draw_fig_3B(data)

draw_fig_3C(data)
}
\arguments{
\item{data}{a \code{data.frame} or a \code{list} containing the data to be plotted}
}
\value{
a ggplot
}
\description{
Functions producing the figures
}
\section{Functions}{
\itemize{
\item \code{draw_fig_1A}: draw fig. 1A

\item \code{draw_fig_1B}: draw fig. 1B

\item \code{draw_fig_1C}: draw fig. 1C

\item \code{draw_fig_1D}: draw fig. 1D

\item \code{draw_fig_2}: draw fig. 2

\item \code{draw_fig_3A}: draw fig. 3A

\item \code{draw_fig_3B}: draw fig. 3B

\item \code{draw_fig_3C}: draw fig. 3C
}}

\examples{
# See ?twinR

}
