% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_models}
\alias{fit_models}
\alias{fit_mod_delivery}
\title{Fit the models}
\usage{
fit_mod_delivery(
  mother_level_data = NULL,
  args_spaMM = list(),
  verbose = FALSE
)
}
\arguments{
\item{mother_level_data}{a \code{tibble} or \code{data.frame} with mother level data}

\item{args_spaMM}{list of additional arguments to pass to the function \code{\link[spaMM:fitme]{fitme}}}

\item{verbose}{whether to display the formula of the fit during the fitting procedure}
}
\value{
the fitted model
}
\description{
These functions fit the models used in the paper.
They have been programmed to be used with our specific data structure.
}
\section{Functions}{
\itemize{
\item \code{fit_mod_delivery}: Fit the model predicting the total number of births per mother from her twinning status
}}

