% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_models}
\alias{fit_models}
\alias{fit_totalbirths}
\alias{fit_twinner.allbirths}
\alias{fit_twinner.firstbirth}
\alias{fit_twinning.binomial}
\alias{fit_AFB}
\alias{fit_PP}
\alias{fit_IBI}
\alias{fit_twinning.binary}
\alias{fit_life_histories}
\alias{compute_slope_from_birth.level.data}
\title{Fit the models}
\usage{
fit_totalbirths(
  mother_level_data,
  when_twinner = "allbirths",
  args_spaMM = list(),
  verbose = TRUE
)

fit_twinner.allbirths(mother_level_data, args_spaMM = list(), verbose = TRUE)

fit_twinner.firstbirth(mother_level_data, args_spaMM = list(), verbose = TRUE)

fit_twinning.binomial(mother_level_data, args_spaMM = list(), verbose = TRUE)

fit_AFB(mother_level_data, args_spaMM = list(), verbose = TRUE)

fit_PP(
  birth_level_data,
  poly_order = NA,
  twin_as.predictor = TRUE,
  args_spaMM = list(),
  verbose = TRUE
)

fit_IBI(
  birth_level_data,
  poly_order = NA,
  twin_as.predictor = TRUE,
  args_spaMM = list(),
  verbose = TRUE
)

fit_twinning.binary(
  birth_level_data,
  poly_order = NA,
  maternal_ID_as.predictor = TRUE,
  args_spaMM = list(),
  verbose = TRUE
)

fit_life_histories(
  scenario,
  birth_level_data,
  args_spaMM = list(),
  verbose = TRUE
)

compute_slope_from_birth.level.data(birth_level_data, verbose = TRUE)
}
\arguments{
\item{mother_level_data}{a \code{tibble} or \code{data.frame} with mother level data}

\item{when_twinner}{a string of characters indicating if the twinning status is based on all births ('allbirths') or just on the first one ('firstbirth')}

\item{args_spaMM}{list of additional arguments to pass to the function \code{\link[spaMM:fitme]{fitme}}}

\item{verbose}{whether to display the formula of the fit during the fitting procedure}

\item{birth_level_data}{a \code{tibble} or \code{data.frame} with (expanded) birth level data}

\item{poly_order}{an integer value defining the polynomial order when considering the effect of age and parity (default = \code{NA}, find best value between 0 and 6)}

\item{twin_as.predictor}{whether to include the variable \code{twin} as a predictor or not in some models (default = \code{TRUE})}

\item{maternal_ID_as.predictor}{whether to include the variable \code{maternal_ID} as a random effect predictor or not in some models (default = \code{TRUE})}

\item{scenario}{the scenario defining which models to be fitted: e.g. "ABCD", "AC"... (see paper for explanations)}
}
\value{
the fitted model
}
\description{
These functions fit the models used in the paper.
They have been programmed to be used with our specific data structure.
See \strong{Functions}, below, for details on each function.
}
\section{Functions}{
\itemize{
\item \code{fit_totalbirths}: fit the model predicting the total number of births per mother from her twinning status

\item \code{fit_twinner.allbirths}: fit the model predicting the twinning status of the mother at any birth during her life from her total number of births

\item \code{fit_twinner.firstbirth}: fit the model predicting the twinning status of the mother at first birth from her total number of births

\item \code{fit_twinning.binomial}: fit the model predicting the probability of a birth to result in twins from the total number of births

\item \code{fit_AFB}: fit the model predicting the age at first birth from the twinning status and the total number of births

\item \code{fit_PP}: fit the model predicting the probability of parity progression

\item \code{fit_IBI}: fit the model predicting the duration of the interbirth interval (minus 6 months)

\item \code{fit_twinning.binary}: fit the model predicting the probability of twinning for a given birth event

\item \code{fit_life_histories}: fit all three life history models according to the simulation scenario

This function fits all three life history models according to the simulation scenario provided.
The polynomial orders are estimated, which is why each of the three models has to be fitted seven
times (order 0->6). It also returns the total time elapsed by the function.

\item \code{compute_slope_from_birth.level.data}: compute the slope between the total number of births and the per-birth twinning probability from birth level data

This function fits the model investigating the relationship between parity and twinning probability using \code{\link{fit_twinning.binomial}} and retrieve the slope of interest.
For this, it aggregates the birth level data.
}}

\examples{
# See ?twinR

}
