% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{#}
\alias{#}
\title{Show an object as text output}
\usage{
`#`(result, digits = 0, prefix = "#' #")
}
\arguments{
\item{result}{an object to output}

\item{digits}{the number of digits to display (default = 0)}

\item{prefix}{the prefix added to each line (the default is roxygen friendly)}
}
\description{
This function makes copying and pasting the results of the analyses more straightforward.
It adds a pound sign (\verb{#}), or equivalent, before displaying each line of the output.
}
\examples{
test <- runif(3)
test
round(test, digits = 2)
`#`(test, digits = 2)
`#`(test, prefix = "#")
`#`(data.frame(x = test, y = test), digits = 2, prefix = "#")

\dontrun{
## do not attempt to round a list:
  `#`(list(x = test, y = test), digits = 2, prefix = "#")
}

`#`(list(x = test, y = test), prefix = "#")

}
