% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit.R
\name{simulate_slopes}
\alias{simulate_slopes}
\title{Simulate slopes under a given scenario}
\usage{
simulate_slopes(
  birth_level_data,
  scenario,
  life_history_fits = NULL,
  seed = 123L,
  N_replicates = 49L,
  args_spaMM = list(),
  verbose = list(fit = FALSE, simu = FALSE)
)
}
\arguments{
\item{birth_level_data}{a \code{tibble} or \code{data.frame} with (expanded) birth level data}

\item{scenario}{the scenario defining which models to be fitted: e.g. "ABCD", "AC"... (see paper for explanations)}

\item{life_history_fits}{a named list containing the fits of each three life history models}

\item{seed}{an \code{integer} providing the seed for the random generator}

\item{N_replicates}{the number of simulation replicates to run}

\item{args_spaMM}{list of additional arguments to pass to the function \code{\link[spaMM:fitme]{fitme}}}

\item{verbose}{whether to display the formula of the fit during the fitting procedure}
}
\value{
a list containing the simulated slopes, the input seed, and the time elapsed to do the job
}
\description{
This functions allows for the production of slopes of interest (the one quantifying effect of
total_birth on the log odd of twinning) taking a simulation scenario as true. This thus
corresponds to the computation of the slope under the null hypothesis (see Details for more
information).
}
\details{
The function runs a first simulation according to the life history models fitted on the observed
data (level 1). Then, the three life history models are refitted on the simulated data that has
been produced and these new fits are used to simulated new data (level 2). The slope of interest
is then computed using each of the simulated dataset from level 2.
}
\examples{
## See ?twinR

}
\seealso{
run_simulation
}
